backend/requirements.txt
supabase
python-dotenv
flask
flask_cors
redis